\name{beetles}
\alias{beetles}
\docType{data}
\title{Mortality of confused flour beetles}
\description{
  Data from a study examining the response of confused flour beetles
  to increasing concentrations of gaseous carbon disulphide. After
  exposure for five hours the exact concentration of carbon disulphide
  was determined and the number of dead flour beetles were recorded. For
  each concentration of carbon disulphide duplicate batches of beetles
  were used. 
}
\usage{data(beetles)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{conc}}{Concentration of carbon disulphide (mg/l)}
    \item{\code{rep}}{Replicate number}
    \item{\code{y}}{Number of deaths in each dose gruop}
    \item{\code{n}}{Total number of beetles in each dose group}
  }
}
\source{
  Strand, A.L. (1930) Measuring the toxicity of insect
  fumigants. \emph{Industrial and Engineering Chemistry: analytical
  edition}, \bold{83}, 426-431. 
}
\references{
  Collett, D. (2003) \emph{Modelling binary data}. Second
  edition. Chapman & Hall/CRC.   
}
\keyword{datasets}
