\name{serum}
\alias{serum}
\docType{data}
\title{Anti-pneumococcus serum}
\description{
  Data from a study examining the protective effect of a serum
  co-administered in increasing doses with an infecting dose of a
  pneumococci culture. Each dose group consisted of 40 mice (\eqn{n})
  and the number of deaths caused by pneumonia was recorded (\eqn{y}) 
}
\usage{data(serum)}
\format{
  A data frame with 5 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{Dose of the anti-pneumococcus serum administered}
    \item{\code{y}}{Number of deaths in each dose group}
    \item{\code{n}}{Total number of mice in each dose group}
  }
}
\source{
  Smith, W. (1932) The titration of antipneumococcus
  serum. \emph{Journal of Pathology}, \bold{35}, 509-526.
}
\references{
  Collett, D. (2003) \emph{Modelling binary data}. Second
  edition. Chapman & Hall/CRC.   
}
\keyword{datasets}
